Conflicts:
	src/Main.java
	src/PageAccueil.java

+++ src/Etudiant.java
@@ -1,5 +1,6 @@
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.net.URL;
 import java.net.URLConnection;
@@ -29,20 +30,30 @@ public class Etudiant {
 		this.present=false;
 	}
 	
+	
+	/**
+	 * Télécharge la photo d'un étudiant
+	 * @return Booléen : true si la photo a été téléchargée ou existe déjà, false sinon
+	 */
 	public boolean telechargerPhoto() {
 		try {
 			URL adresse = new URL(this.lienPhoto);
-			URLConnection con = adresse.openConnection();
-			BufferedInputStream in = new BufferedInputStream(con.getInputStream());
-			BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream("img/" + this.numeroEtudiant + ".jpg"));
-			byte[] read = new byte[1];
-			int len = 1;
-			while ((len = in.read(read)) > 0) {
-				out.write(read, 0, len);
+			File photo = new File("img/" + this.numeroEtudiant + ".jpg");
+			
+			if(!photo.exists()) {
+				URLConnection con = adresse.openConnection();
+				BufferedInputStream in = new BufferedInputStream(con.getInputStream());
+				BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream("img/" + this.numeroEtudiant + ".jpg"));
+				byte[] read = new byte[1];
+				int len = 1;
+				while ((len = in.read(read)) > 0) {
+					out.write(read, 0, len);
+				}
+				out.flush();
+				out.close();
+				in.close();
 			}
-			out.flush();
-			out.close();
-			in.close();
+			
 			return true;
 		}
 		catch (Exception e) {
